+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+**27.2.2026 01.47** Fix WASM crash when a WebDAV client connects.
+  - Move HTTP parsing buffers off the WASM stack and into persistent service storage.
+  - Log accepted connections and request method/path to narrow down future failures.
+
 **30.1.2026 15.25** This is an example changelog entry.
   - First line contains date and time and short description of the change.
   - Additional information can be given as a bulleted list.